backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.17.0
packaging==25.0
requests==2.32.3
SQLAlchemy==2.0.41
transbank-sdk @ git+https://github.com/TransbankDevelopers/transbank-sdk-python.git@a937bfcedaf27be06ab8f675f6c8e8ba39842aa5
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
